completed validations, updated create, update, and destroy controller…

+++ app/controllers/cats_controller.rb
@@ -6,7 +6,11 @@ class CatsController < ApplicationController
 
     def create
         cat = Cat.create(cat_params)
-        render json: cat
+        if cat.valid?
+            render json: cat
+        else
+            render json: cat.errors, status: 422
+        end
     end
 
     def update
@@ -15,13 +19,17 @@ class CatsController < ApplicationController
         if cat.valid?
             render json: cat
         else
-            render json: cat.errors
+            render json: cat.errors, status: 422
         end
     end
 
     def destroy
         cat = Cat.find(params[:id])
-        cat.destroy
+        if cat.valid?
+            cat.destroy
+        else
+            render json: cat.errors, status: 422
+        end
     end
 
     private
